refactor : emergencyMarker 로직 커스텀 훅으로 빼기
